Moved Transport trait and auto-implemented for references of types implementing it
