+++ agents/rust/aries-vcx-agent/src/http_client.rs
@@ -1,6 +1,5 @@
-use aries_vcx::{
-    agency_client::httpclient::post_message, errors::error::VcxResult, protocols::connection::Transport,
-};
+use aries_vcx::{agency_client::httpclient::post_message, errors::error::VcxResult, transport::Transport};
+
 use async_trait::async_trait;
 
 pub struct HttpClient;
