+++ aries_vcx/src/lib.rs
@@ -49,6 +49,7 @@ pub mod common;
 pub mod core;
 pub mod errors;
 pub mod plugins;
+pub mod transport;
 
 #[cfg(test)]
 pub mod test {
