+++ aries_vcx/src/protocols/connection/common/states/responded.rs
@@ -1,6 +1,6 @@
 use messages::diddoc::aries::diddoc::AriesDidDoc;
 
-use crate::protocols::connection::traits::{TheirDidDoc, ThreadId};
+use crate::protocols::connection::trait_bounds::{TheirDidDoc, ThreadId};
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub struct RespondedState {
